cleanup and config updates
