--- cbws/conf/ws_server_config_schema.json
@@ -1,29 +1,0 @@
-{
-  "$schema": "http://json-schema.org/schema#",
-  "id": "/conf/module_configs_schema",
-  "description": "websocket server configuration.",
-  "type": "object",
-  "additionalProperties": false,
-  "properties": {
-    "ws_server_port": {
-      "description": "",
-      "type": "integer",
-      "required": true
-    },
-    "rabbitmq_address": {
-      "description": "RabbitMQ IP address",
-      "type": "string",
-      "required": true
-    },
-    "rabbitmq_user": {
-      "description": "RabbitMQ user",
-      "type": "string",
-      "required": true
-    },
-    "rabbitmq_pwd": {
-      "description": "RabbitMQ password",
-      "type": "string",
-      "required": true
-    }
-  }
-}
