+++ examples/ws_server_config.docker.json
